% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setup_js_html}
\alias{setup_js_html}
\title{Sets up the html and javascript scripts in the current folder for visualization.}
\usage{
setup_js_html()
}
\value{
Does not return anything.
}
\description{
Sets up the html and javascript scripts in the current folder for visualization.
}
\details{
Copies inst/scripts and inst/viz.html from package to the current folder.
Then inserts subfolder names under "dats/" that contain "cors.json", "dat.json" and "graphs.json" in viz.html between the lines "folder_names = [" and "]".
}
\examples{
setup_js_html()
}
