% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{viz}
\alias{viz}
\title{Main function for estimating and writing self/differential correlation matrices to local files.}
\usage{
viz(
  dat_name,
  dat1X,
  dat2X,
  dat1Y = NULL,
  dat2Y = NULL,
  name_dat1 = "1",
  name_dat2 = "2",
  cor_names = c("pearson", "kendall", "spearman", "nonparakendall", "nonparaspearman"),
  permutation = TRUE,
  alpha = 0.05,
  sides = 2,
  B = 1000,
  adj_method = "BY",
  parallel = FALSE,
  verbose = TRUE,
  make_plot = TRUE,
  perm_seed = NULL,
  Cai_seed = NULL,
  layout_seed = NULL
)
}
\arguments{
\item{dat_name}{A string, a given name for this dataset. Files for visualization will be saved under \code{file.path("dats", dat_name)}.}

\item{dat1X}{A matrix data for group X for the first sample; see details. Must not be \code{NULL} and must have the same number of columns as \code{dat2X}.}

\item{dat2X}{A matrix data for group X for the second sample; see details. Must not be \code{NULL} and must have the same number of columns as \code{dat1X}.}

\item{dat1Y}{Optional, a matrix data for group Y for the first sample and defaults to \code{NULL}; see details. If not \code{NULL}, must have the same number of rows as \code{dat1X} and same number of columns as \code{dat2Y}, and \code{dat2Y} must not be \code{NULL}.}

\item{dat2Y}{Optional, a matrix data for group X for the second sample and defaults to \code{NULL}; see details. If not \code{NULL}, must have the same number of rows as \code{dat2X} and same number of columns as \code{dat1Y}, and \code{dat1Y} must not be \code{NULL}.}

\item{name_dat1}{A string, name for the first sample. Defaults to "1".}

\item{name_dat2}{A string, name for the second sample. Defaults to "2".}

\item{cor_names}{A string or a vector of strings, name(s) of correlation types to be estimated. Must be chosen from \code{"pearson"}, \code{"kendall"}, \code{"spearman"}, \code{"nonparakendall"}, and \code{"nonparaspearman"}.}

\item{permutation}{Logical, indicating whether permutation tests should be done in addition to parametric tests; defaults to \code{TRUE}.}

\item{alpha}{Numerical, the significance level in hypothesis testing; defaults to 0.05.}

\item{sides}{A number \code{1}, \code{2}, \code{3} or a matrix containing \code{1}, \code{2}, \code{3}. If a matrix, must be of size \code{ncol(dat1X) x ncol(dat1X)} if \code{dat1Y} is \code{NULL}, or \code{ncol(dat1X) x ncol(dat1Y)} otherwise. \code{2} stands for two-sided tests, \code{1} for one-sided test with null hypothesis being the corresponding entries >= 0 (the corresponding correlation for sample 1 stronger than that for sample 2), and \code{3} for one-sided test with null hypothesis being the corresponding entries <= 0.}

\item{B}{An integer, the number of bootstrapping samples in permutation tests; defaults to 1000.}

\item{adj_method}{A string, the method passed to \code{stats::p.adjust} for adjusting the p values for multiple testing; defaults to "BY". Must be one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", or "none".}

\item{parallel}{A logical, whether to use parallel computing; may fail sometimes for some systems and defaults to \code{FALSE}.}

\item{verbose}{A logical, whether to print progress; defaults to \code{TRUE}.}

\item{make_plot}{A logical, whether to make heat maps and static graphs; defaults to \code{TRUE}. Plots will be made under \code{file.path("plots", dat_name)}.}

\item{perm_seed}{A number, seed for permutation test; defaults to \code{NULL}.}

\item{Cai_seed}{A number, seed for the method by Cai and Zhang; defaults to \code{NULL}.}

\item{layout_seed}{A number, seed for the layout of the static graphs; defaults to \code{NULL}.}
}
\value{
Does not return anything, but instead creates folders and files under \code{file.path("dats", dat_name)} and \code{file.path("plots", dat_name)}.
}
\description{
Main function for estimating and writing self/differential correlation matrices to local files.
}
\details{
If \code{dat1Y} and \code{dat2Y} are \code{NULL}, the function estimates the difference \code{cor(dat1X) - cor(dat2X)} and truncates to 0 the entries that are below a certain threshold determined by parameteric/permutation tests.
If \code{dat1Y} and \code{dat2Y} are not \code{NULL}, the difference \code{cor(dat1X, dat1Y) - cor(dat2X, dat2Y)} is estimated.
The dimensions must be as follows: \code{dat1X} has dimension n1 x pX, \code{dat2X} n2 x pX, and if provided, \code{dat1Y} n1 x pY and \code{dat2Y} n2 x pY.
}
\examples{
dat0 <- read.csv(file.path(path.package("CorDiffViz"), "extdata/sample_data.csv"))
# First column of dat0 is the group (dat1 or dat2)
dat1 <- dat0[dat0$Group=="AA", 2:10][1:13,] # 13 x 9
dat2 <- dat0[dat0$Group=="BB", 2:10][1:15,] # 15 x 9
# Self correlations
viz(dat_name="exmp_self", dat1X=dat1, dat2X=dat2, dat1Y=NULL, dat2Y=NULL,
    name_dat1="AA", name_dat2="BB", 
    cor_names=c("pearson","spearman", "kendall","nonparaspearman","nonparakendall"),
    permutation=TRUE, alpha=0.05, sides=2, B=1000, adj_method="BY", verbose=TRUE,
    make_plot=TRUE, parallel=FALSE, perm_seed=1, Cai_seed=1, layout_seed=1)
# Correlations between variables in group X = {1:4} and variables in group Y = {5:9}
viz(dat_name="exmp_XY", dat1X=dat1[,1:(ncol(dat1)/2)], dat2X=dat2[,1:(ncol(dat1)/2)], 
    dat1Y=dat1[,(ncol(dat1)/2+1):ncol(dat1)], dat2Y=dat2[,(ncol(dat1)/2+1):ncol(dat1)], 
    name_dat1="AA", name_dat2="BB", 
    cor_names=c("pearson","spearman", "kendall","nonparaspearman","nonparakendall"), 
    permutation=TRUE, alpha=0.05, sides=2, B=1000, adj_method="BY", verbose=TRUE, 
    make_plot=TRUE, parallel=FALSE, perm_seed=1, layout_seed=1)
    
# Remove folders for the examples generated above
unlink(c("dats/exmp_self", "dats/exmp_XY", "plots/exmp_self", "plots/exmp_XY"), recursive=TRUE)
setup_js_html()
}
